changed quicksort algorithm to the version from cracking the coding interview book. increased warning level in project, set warnings as errors, fixed warnings

+++ RevisionVS2015/CrackingTheCodingInterviewCPlusPlus.h
@@ -157,7 +157,7 @@ void* AlignedMalloc(int size, unsigned int alignment)
 	}
 
 	char* aligned = allocated + beginPointerSize;
-	unsigned int remainder = (unsigned int)aligned % alignment;
+	unsigned int remainder = (unsigned long long)aligned % alignment;
 	if (remainder != 0) {
 		aligned += (alignment - remainder);
 	}
